Backend/requirements.txt
# NeuroNex Backend - Phase 3 (Hackathon)
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
# AI & Machine Learning
google-genai
openai-whisper
scikit-learn
pandas
numpy
scipy
gTTS
pydub

# Utilities
python-dotenv
python-multipart
asyncpg

